small refactoring: deeper file structure
