+++ src/model/Base/string_kernel.py
