+++ src/model/Gnofix/figures/diagram.jpg
