+++ src/model/Gnofix/figures/diagram_resized.png
