+++ src/model/Smooth/cnn.py
